num steps int

+++ emmet-core/emmet/core/electrode.py
@@ -206,7 +206,7 @@ class ConversionElectrodeDoc(ConversionVoltagePairDoc):
         description="The working ion as an Element object",
     )
 
-    num_steps: float = Field(
+    num_steps: int = Field(
         None,
         description="The number of distinct voltage steps in from fully charge to "
         "discharge based on the stable intermediate states",
